Use size_t instead of int

+++ system/lib/pthread/pthread_self_stub.c
@@ -33,6 +33,6 @@ static void init_pthread_self(void) {
   __main_pthread.locale = &libc.global_locale;
   __main_pthread.tid = getpid();
   __main_pthread.stack = &__stack_high;
-  __main_pthread.stack_size = ((int)&__stack_high) - ((int)&__stack_low);
+  __main_pthread.stack_size = ((size_t)&__stack_high) - ((size_t)&__stack_low);
   __main_pthread.guard_size = __default_guardsize;
 }
